Version 24.06

+++ script_js/setfilename.js
@@ -2,13 +2,25 @@ jQuery(document).ready(function(){
 	$("#btn_filename").on('click',function(){
 		sessionStorage.clear();
 
-		$filename = "XML_Files/" + $("#input_filename").val().replace(/C:\\fakepath\\/,"");
-		alert($filename);
+		$filename = $("#input_filename").val();
+
 		if($filename == null || $filename == ""){
-			alert("File non trovato!");
+			//swal("File non trovato!");
+			sweetAlert("Oops...", "Something went wrong!", "error");
 		}else{
+			$filename = "XML_Files/" + $("#input_filename").val().replace(/C:\\fakepath\\/,"");
 			console.log("No Null!");
-			sessionStorage.setItem("filename",$filename);
+			swal({
+				title: "Good job!",
+  				text: ". . File Loaded . .",
+  				type: "success",
+			},function(){
+				sessionStorage.setItem("filename",$filename);
+				  setTimeout(function(){
+				  	console.log("ciao");
+				  	window.location.href = 'home.html';
+				  }, 500);
+				});
 		}
 	});
 });
